Add recommended assets section
Under the add asset section, there is a new list
of all referenced assets that have not already
been attached to the ticket. When clicked, the
asset's name will be appended to the asset name
input box.

Also made the asset name regex more restrictive
to not include netids, only the common computer
prefixes we use.